// web/src/lib/types.ts
export interface RawMaterial {
  id: number;
  store_id: number;
  name: string;
  sku: string;
  unit: string;
  cost_per_unit: number;
  current_stock: number;
}